+++ src/models/Settings.php
@@ -31,7 +31,26 @@ class Settings extends Model
     {
         return [
             [['colors', 'palettes'], 'required'],
-            [['colors', 'palettes'], 'array'],
+            [['colors', 'palettes'], function ($attribute, $params) {
+                if (!is_array($this->colors)) {
+                    $this->addError('colors', Craft::t('colour-swatches', 'colors is not array!'));
+                }
+
+                if (!is_array($this->palettes)) {
+                    $this->addError('palettes', Craft::t('colour-swatches', 'palettes is not array!'));
+                }
+            }],
+            [['palettes'], 'filter', 'filter' => function ($palettes) {
+                foreach ($palettes as &$palette) {
+                    foreach ($palette as &$color) {
+                        if (is_string($color['color'])) {
+                            $color['color'] = json_decode($color['color'], true);
+                        }
+                    }
+                }
+                
+                return $palettes;
+            }]
         ];
     }
 }